:bug: update run script

+++ main.sh
@@ -1,30 +1,40 @@
 #!/usr/bin/env bash
-src="louvain-openmp-adjust-schedule"
-out="/home/resources/Documents/subhajit/$src.log"
+src="louvain-communities-openmp"
+out="$HOME/Logs/$src$1.log"
 ulimit -s unlimited
 printf "" > "$out"
 
 # Download program
-rm -rf $src
-git clone https://github.com/puzzlef/$src
-cd $src
+if [[ "$DOWNLOAD" != "0" ]]; then
+  rm -rf $src
+  git clone https://github.com/puzzlef/$src
+  cd $src
+fi
+
+# Fixed config
+: "${TYPE:=float}"
+: "${MAX_THREADS:=64}"
+: "${REPEAT_METHOD:=5}"
+# Define macros (dont forget to add here)
+DEFINES=(""
+"-DTYPE=$TYPE"
+"-DMAX_THREADS=$MAX_THREADS"
+"-DREPEAT_METHOD=$REPEAT_METHOD"
+)
 
 # Run
-g++ -std=c++17 -O3 -fopenmp main.cxx
-stdbuf --output=L ./a.out ~/data/web-Stanford.mtx      2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/web-BerkStan.mtx      2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/web-Google.mtx        2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/web-NotreDame.mtx     2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/soc-Slashdot0811.mtx  2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/soc-Slashdot0902.mtx  2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/soc-Epinions1.mtx     2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/coAuthorsDBLP.mtx     2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/coAuthorsCiteseer.mtx 2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/soc-LiveJournal1.mtx  2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/coPapersCiteseer.mtx  2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/coPapersDBLP.mtx      2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/indochina-2004.mtx    2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/italy_osm.mtx         2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/great-britain_osm.mtx 2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/germany_osm.mtx       2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/asia_osm.mtx          2>&1 | tee -a "$out"
+g++ ${DEFINES[*]} -std=c++17 -O3 -fopenmp main.cxx
+# stdbuf --output=L ./a.out ~/Data/web-Stanford.mtx   0 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/indochina-2004.mtx  0 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/uk-2002.mtx         0 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/arabic-2005.mtx     0 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/uk-2005.mtx         0 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/webbase-2001.mtx    0 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/it-2004.mtx         0 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/sk-2005.mtx         0 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/com-LiveJournal.mtx 1 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/com-Orkut.mtx       1 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/asia_osm.mtx        1 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/europe_osm.mtx      1 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/kmer_A2a.mtx        1 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/kmer_V1r.mtx        1 0 2>&1 | tee -a "$out"
